+++ Projects/Features/SignUp/Sources/Core/AppDelegate.swift
@@ -1,19 +1,31 @@
-import Foundation
+//
+//  AppDelegate.swift
+//  SignUp
+//
+//  Created by mincheol on 2022/09/11.
+//  Copyright © 2022 cocaine.io. All rights reserved.
+//
+
 import UIKit
-import SignUp
 
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
+@main class AppDelegate: UIResponder, UIApplicationDelegate {
+    
     var window: UIWindow?
-
-    internal func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        let vc = UIViewController()
-        vc.title = "SignUp"
-        vc.view.backgroundColor = .systemYellow
-        self.window = UIWindow(frame: UIScreen.main.bounds)
-        self.window?.rootViewController = UINavigationController(rootViewController: vc)
-        self.window?.makeKeyAndVisible()
-
+    
+    func application(
+        _ application: UIApplication,
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil
+    ) -> Bool {
+        window = UIWindow(frame: UIScreen.main.bounds)
+        let viewController = SignUpViewController.create()
+        
+        // TODO: 나중에 바꾸기
+        let rootViewController =  UINavigationController(rootViewController: viewController)
+        window?.rootViewController = rootViewController
+        
+//        window?.rootViewController = viewController
+        window?.makeKeyAndVisible()
         return true
     }
+    
 }
