--- Projects/Features/SignUp/Sources/Data/SignUp.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct SignUp {
-    public init() {}
-}
