--- Projects/Features/SignUp/Sources/Presentation/SignUp.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct SignUp {
-    public init() {}
-}
